+++ backend/unithis/src/main/java/com/unithis/service/ChatRoomService.java
@@ -17,13 +17,13 @@ public class ChatRoomService implements IChatRoomService {
 	private final ChatRoomMapper chatroomMapper;
 
 	@Override
-	public int createChatRoom(ChatRoom newRoom) {
-		int idx = chatroomMapper.createChatRoom(newRoom);
+	public long createChatRoom(ChatRoom newRoom) {
+		long idx = chatroomMapper.createChatRoom(newRoom);
 		return idx == 1 ? newRoom.getId() : -1;
 	}
 
 	@Override
-	public List<ChatRoom> getJoinedRoomList(int user_id) {
+	public List<ChatRoom> getJoinedRoomList(long user_id) {
 		return chatroomMapper.getJoinedRoomList(user_id);
 	}
